+++ packages/autodoc.lua
@@ -1,0 +1,8 @@
+SILE.registerCommand("package-documentation", function (o,c)
+  local package = SU.required(o, "src", "src for package documentation")
+  SILE.process(
+    SILE.inputs.TeXlike.docToTree(
+      require(package).documentation
+    )
+  )
+end)